Bug fixes and refactoring of landmark mapper
