--- path-finding/cone_mapper/__init__.py
@@ -1,1 +1,0 @@
-# cone_mapper
